# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"database/sql"

	"github.com/google/uuid"
	"github.com/sqlc-dev/pqtype"
)

type CustomClaim struct {
	UserID     int32                 `json:"user_id"`
	ClaimKey   string                `json:"claim_key"`
	ClaimValue pqtype.NullRawMessage `json:"claim_value"`
	CreatedAt  sql.NullTime          `json:"created_at"`
	UpdatedAt  sql.NullTime          `json:"updated_at"`
}

type Role struct {
	ID          int32          `json:"id"`
	Name        string         `json:"name"`
	Description sql.NullString `json:"description"`
	CreatedAt   sql.NullTime   `json:"created_at"`
}

type SecurityLog struct {
	ID               int32          `json:"id"`
	UserID           int32          `json:"user_id"`
	EventType        string         `json:"event_type"`
	EventDescription string         `json:"event_description"`
	IpAddress        pqtype.Inet    `json:"ip_address"`
	UserAgent        sql.NullString `json:"user_agent"`
	CreatedAt        sql.NullTime   `json:"created_at"`
}

type Session struct {
	ID         uuid.UUID      `json:"id"`
	UserID     int32          `json:"user_id"`
	Device     string         `json:"device"`
	Location   sql.NullString `json:"location"`
	IpAddress  pqtype.Inet    `json:"ip_address"`
	UserAgent  sql.NullString `json:"user_agent"`
	LastActive sql.NullTime   `json:"last_active"`
	CreatedAt  sql.NullTime   `json:"created_at"`
	RevokedAt  sql.NullTime   `json:"revoked_at"`
}

type User struct {
	ID                  int32          `json:"id"`
	Auth0ID             string         `json:"auth0_id"`
	Name                string         `json:"name"`
	Email               string         `json:"email"`
	Picture             sql.NullString `json:"picture"`
	MfaEnabled          sql.NullBool   `json:"mfa_enabled"`
	CreatedAt           sql.NullTime   `json:"created_at"`
	UpdatedAt           sql.NullTime   `json:"updated_at"`
	PasswordlessEnabled sql.NullBool   `json:"passwordless_enabled"`
}

type UserRole struct {
	UserID    int32        `json:"user_id"`
	RoleID    int32        `json:"role_id"`
	GrantedBy int32        `json:"granted_by"`
	GrantedAt sql.NullTime `json:"granted_at"`
}